+++ uninstall.php
@@ -18,6 +18,7 @@ static $options_name = array(
 	'editor_emoji',
 	'editor_toc',
 	'editor_katex',
+	'editor_mermaid',
 	'editor_flow',
 	'editor_mindmap',
 	'editor_sequence',